requirements.txt
pydantic~=2.9.2
requests~=2.32.3
python-telegram-bot~=21.6
urllib3~=2.2.3
beautifulsoup4~=4.12.3